Move MonitoredInputFile and split out nested classes
